BUG: Ensure valid range of bin edges
When using IBSI binning, a custom lower edge is used if a resegmentation range is indicated. While this will always result a valid range of bin edges for the original image, it is possible that there are gray values below this range when using filters.
np.discretize will assign "0" to those gray values, which will result in faillure to calculate the texture matrix. Therefore prevent an invalid lower bin edge and log a warning if using this escape.